Import INTEL 20030123 release.

+++ sys/dev/acpi/acpica/Subsystem/acoutput.h
@@ -1,7 +1,7 @@
 /******************************************************************************
  *
  * Name: acoutput.h -- debug output
- *       $Revision: 1.1.1.3 $
+ *       $Revision: 1.1.1.4 $
  *
  *****************************************************************************/
 
@@ -9,7 +9,7 @@
  *
  * 1. Copyright Notice
  *
- * Some or all of this work - Copyright (c) 1999 - 2002, Intel Corp.
+ * Some or all of this work - Copyright (c) 1999 - 2003, Intel Corp.
  * All rights reserved.
  *
  * 2. License
@@ -209,7 +209,7 @@
 /*
  * Debug level macros that are used in the DEBUG_PRINT macros
  */
-#define ACPI_DEBUG_LEVEL(dl)       dl,__LINE__,&_Dbg
+#define ACPI_DEBUG_LEVEL(dl)        (UINT32) dl,__LINE__,&_Dbg
 
 /* Exception level -- used in the global "DebugLevel" */
